+++ src/test/java/api/stepdefinitions/NewUser.java
@@ -5,8 +5,8 @@ import api.actions.UtilActions;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import ui.context.ObjectKeys;
-import ui.context.ScenarioContext;
+import context.ObjectKeys;
+import context.ScenarioContext;
 
 import static org.junit.Assert.assertEquals;
