+++ src/test/java/cfg/BrowserDriver.java
@@ -1,4 +1,4 @@
-package ui.cfg;
+package cfg;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -9,7 +9,7 @@ import org.openqa.selenium.firefox.FirefoxOptions;
 import org.openqa.selenium.edge.EdgeOptions;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
-import ui.utils.ConfigReader;
+import utililities.ConfigReader;
 
 public class BrowserDriver {
     private static final String BROWSER_TYPE = new ConfigReader().getProperty("browserType").trim().toLowerCase();
