+++ src/test/java/runners/TestRunnerTwoInOne.java
@@ -7,7 +7,7 @@ import io.cucumber.junit.Cucumber;
 @RunWith(Cucumber.class)
 @CucumberOptions(
         features = {"./src/test/resources/features/api", "./src/test/resources/features/ui"},
-        glue = {"api/stepdefinitions", "ui/stepdefinitions", "ui/cfg", "ui/context"},
+        glue = {"api/stepdefinitions", "ui/stepdefinitions", "cfg", "context"},
         plugin = {"pretty", "html:target/cucumber-html-report", "json:cucumber.json"},
         tags = "@Run"
 )
