+++ src/test/java/runners/TestRunnerUI.java
@@ -7,7 +7,7 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(
         features = "./src/test/resources/features/ui",
-        glue = {"ui/stepdefinitions", "ui/cfg", "ui/context"},
+        glue = {"ui/stepdefinitions", "cfg", "context"},
         plugin = {"pretty", "html:target/cucumber-html-report","json:cucumber.json"},
         tags = "@Run"
 )
