+++ src/test/java/utililities/ConfigReader.java
@@ -1,4 +1,4 @@
-package ui.utils;
+package utililities;
 
 import java.io.FileInputStream;
 import java.io.IOException;
